requirements.txt
moderngl
numpy
pygame
PyGLM
pywavefront
pillow
